test: Added tests for the UI
Closes #44

+++ Makefile
@@ -1,5 +1,7 @@
 docker_tag=latest
 all=false
+ui_host=http://localhost:8080
+start_containers=true
 
 clean:
 	./gradlew clean
@@ -17,6 +19,16 @@ docker-run:
 	-p 8080:8080 \
 	devatherock/velocity-template-tester:$(docker_tag)
 integration-test:
-	DOCKER_TAG=$(docker_tag) docker-compose up &
+	DOCKER_TAG=$(docker_tag) docker compose up --wait
 	./gradlew integrationTest
-	docker-compose down	
+	docker-compose down
+ui-test:
+ifeq ($(start_containers), true)
+	DOCKER_TAG=$(docker_tag) docker compose up --wait
+endif	
+	UI_HOST=$(ui_host) npx playwright test
+ifeq ($(start_containers), true)
+	docker-compose down
+endif			
+ui-test-debug:
+	UI_HOST=$(ui_host) npx playwright test --ui --headed
